Review and translate model settings UI
