+++ app/assets/l10n/intl_en.arb
@@ -181,7 +181,11 @@
       }
     }
   },
+  "modelGeneralSettingsLabel": "General",
   "modelIsGeneratingSnackBar": "Model is generating",
+  "modelMiscSettingsLabel": "Miscellaneous",
+  "modelPenaltySettingsLabel": "Penalty",
+  "modelPerformanceSettingsLabel": "Performance",
   "modelPulledSuccessfullySnackBar": "Model pulled successfully",
   "modelPushedSuccessfullySnackBar": "Model pushed successfully",
   "modelsPageCreateButton": "Create",
@@ -199,6 +203,7 @@
   "modelsPagePullButton": "Pull",
   "modelsPagePushButton": "Push",
   "modelsPageRefreshButton": "Refresh",
+  "modelsPageSettingsButton": "Open model settings",
   "modelsPageTitle": "Models management",
   "modelsPageUseButton": "Use model",
   "modifiedAtTextShared": "Modified at: {modifiedAt}",
@@ -275,8 +280,15 @@
   "settingsLanguageItalian": "Italian",
   "settingsLanguageSpanish": "Spanish",
   "settingsLanguageSystem": "System",
-  "settingsModelTitle": "Model specific settings",
   "settingsModelDefault": "Default",
+  "settingsModelDialogTitle": "{model} specific settings",
+  "@settingsModelDialogTitle": {
+    "placeholders": {
+      "model": {
+        "type": "String"
+      }
+    }
+  },
   "settingsModelOff": "On",
   "settingsModelOn": "Off",
   "settingsPageAccessibilityLabel": "Accessibility",
